refactor a few things
