+++ vqls_prototype/solver/log.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+from typing import Optional, Union, List, Callable, Dict, Tuple
+import numpy as np
+
+@dataclass
+class VQLSLog:
+    values: List
+    parameters: List
+
+    def update(self, count, cost, parameters):
+        self.values.append(cost)
+        self.parameters.append(parameters)
+        print(f"VQLS Iteration {count} Cost {cost: .3e}", end="\r", flush=True)
