+++ vqls_prototype/solver/variational_linear_solver.py
@@ -57,6 +57,20 @@ class VariationalLinearSolverResult(VariationalResult):
         """
         self._state = state
 
+    @property
+    def vector(self) -> np.ndarray:
+        """returns the actual solution of the linear system"""
+        return self._vector
+
+    @vector.setter
+    def vector(self, vector: np.ndarray) -> None:
+        """Set the solution vector of the linear system
+
+        Args:
+            vector: The solution vector.
+        """
+        self._vector = vector
+
 
 class VariationalLinearSolver(ABC):
     """An abstract class for linear system solvers in Qiskit."""
